+++ CHANGELOG.md
@@ -14,6 +14,8 @@ This file follows the best practices from [keepachangelog.com](http://keepachang
 
 ### Changed
 
+- Redirect user to the page where user was before login/sign up [#2758](https://github.com/sharetribe/sharetribe/pull/2758)
+
 ### Deprecated
 
 ### Removed